+++ cmd/fireplace-viewer/app/assets/debounce/debounce.js
@@ -1,0 +1,20 @@
+let f = function(fn, delay) {
+	let id = null;
+
+	return function() {
+		let args = arguments;
+		let self = this;
+
+		clearTimeout(id);
+
+		id = setTimeout(function() {
+			fn.apply(self, args);
+		}, delay);
+	};
+};
+
+if (window !== undefined) {
+	window.debounce = f;
+}
+
+export default f;
