+++ cmd/fireplace-viewer/app/assets/fireplace-viewer/css/styles.css
@@ -25,3 +25,69 @@ body {
 .pointer {
 	cursor: pointer;
 }
+
+/******************************************************************************
+ * Table styles
+ *****************************************************************************/
+
+.table-header {
+	top: 56px;
+	position: fixed !important;
+}
+
+.after-header {
+	margin-top: 50px;
+}
+
+.icon-error-status {
+	color: #ef4d3b;
+}
+
+.icon-warning-status {
+	color: #fcfc4b;
+}
+
+.icon-debug-status {
+	color: #33f7e3;
+}
+
+.icon-info-status {
+	color: #0de815;
+}
+
+.column-icon {
+	width: 5%;
+}
+
+.column-message {
+	width: 65%;
+}
+
+.column-application {
+	width: 10%;
+}
+
+.column-date {
+	width: 20%;
+}
+
+section {
+	margin: 5px 0px 20px 0px;
+}
+
+.page-number {
+	padding: 0.5rem 1rem;
+}
+
+.right-panel {
+	background-color: white;
+	border: 1px solid #ccc;
+	height: 100%;
+	width: 450px;
+	position: fixed;
+	right: -460px;
+	top: 56px;
+	padding: 10px 30px 5px;
+	overflow-x: hidden;
+	transition: 0.5s;
+}
