+++ cmd/fireplace-viewer/app/components/click-outside/click-outside.js
@@ -1,7 +1,5 @@
-import Vue from "vue";
-
 Vue.directive("click-outside", {
-	bind: function (el, binding, vnode) {
+	bind(el, binding, vnode) {
 		el.clickOutsideEvent = function (event) {
 			if (!(el == event.target || el.contains(event.target))) {
 				vnode.context[binding.expression](event);
@@ -11,7 +9,7 @@ Vue.directive("click-outside", {
 		document.body.addEventListener("click", el.clickOutsideEvent);
 	},
 
-	unbind: function (el) {
+	unbind(el) {
 		document.body.removeEventListener("click", el.clickOutsideEvent);
 	}
 });
