+++ go.mod
@@ -1,5 +1,7 @@
 module github.com/app-nerds/fireplace
 
+go 1.13
+
 require (
 	github.com/app-nerds/kit/v4 v4.0.0
 	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8
@@ -8,5 +10,3 @@ require (
 	github.com/pkg/errors v0.8.1
 	github.com/sirupsen/logrus v1.4.2
 )
-
-go 1.13
